refactor

+++ neon-note-project/src/components/Task/ViewModel/useTaskProjects.tsx
@@ -6,6 +6,7 @@ import { errorToast, successToast } from '@/utils/toasts/toasts';
 import { useRouter } from 'next/router';
 import { generateIdProjects } from '@/utils/generateId';
 import { v4 as uuidv4 } from 'uuid';
+import { useDisclosure } from '@chakra-ui/react';
 
 export function useTaskProjects() {
   const {
@@ -17,6 +18,12 @@ export function useTaskProjects() {
     setFoldersTask,
   } = useContextTaskData();
 
+  const {
+    open: isOpenModalCreateProject,
+    onOpen: onOpenModalCreateProject,
+    onClose: onCloseModalCreateProject,
+  } = useDisclosure();
+
   const router = useRouter();
   const { id } = router.query;
 
@@ -26,34 +33,54 @@ export function useTaskProjects() {
       return;
     }
 
-    const newProject: ProjectProps = {
-      id: uuidv4(),
-      projectName: projectName,
-      projectTasks: {
-        total: 0,
-        status: {
-          toStart: [],
-          inProgress: [],
-          finished: [],
-        },
-      },
-    };
-
-    const updatedFolders = foldersTask.map(folder =>
-      folder.id === id
-        ? { ...folder, projects: [...folder.projects, newProject] }
-        : folder
+    const projectNameExists = listProjects.some(
+      project => project.projectName === projectName
     );
 
-    if (typeof window !== 'undefined') {
-      localStorage.setItem('foldersTask', JSON.stringify(updatedFolders));
-      setFoldersTask(updatedFolders);
-      setListProjects(prevProjects => [...prevProjects, newProject]);
+    if (!projectName) {
+      errorToast('Nome do projeto deve ter pelo menos 1 caractere.');
+      return;
     }
 
-    setNewTaskProjectName('');
-    successToast('Projeto criado!');
-  };
+    if (projectNameExists) {
+      errorToast(`Projeto "${projectName}" já existe.`);
+      return;
+    }
+
+    try {
+      const newProject: ProjectProps = {
+        id: uuidv4(),
+        projectName: projectName,
+        projectTasks: {
+          total: 0,
+          status: {
+            toStart: [],
+            inProgress: [],
+            finished: [],
+          },
+        },
+      };
+
+      const updatedFolders = foldersTask.map(folder =>
+        folder.id === id
+          ? { ...folder, projects: [...folder.projects, newProject] }
+          : folder
+      );
+
+      if (typeof window !== 'undefined') {
+        localStorage.setItem('foldersTask', JSON.stringify(updatedFolders));
+        setFoldersTask(updatedFolders);
+        setListProjects(prevProjects => [...prevProjects, newProject]);
+      }
+
+      setNewTaskProjectName('');
+      onCloseModalCreateProject();
+      successToast('Projeto criado!');
+    } catch (error) {
+      errorToast('Erro ao criar projeto, tente novamente.');
+      console.error('Erro ao criar projeto:', error);
+    }
+  }
 
   function deleteTaskProject(projectId: string) {
     const updatedFolders = foldersTask.map(folder => ({
@@ -66,8 +93,8 @@ export function useTaskProjects() {
       setFoldersTask(updatedFolders);
       successToast('Projeto deletado!');
       router.replace(`/tasksFolders/${id}`);
-    };
-  };
+    }
+  }
 
   const selectedFolder = foldersTask.find(folder => folder.id === id);
   function getTaskProjects() {
@@ -101,5 +128,8 @@ export function useTaskProjects() {
     listProjects,
     foldersTask,
     deleteTaskProject,
+    isOpenModalCreateProject,
+    onOpenModalCreateProject,
+    onCloseModalCreateProject,
   };
 }
